+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-
+### Fixed
+- ->col() with empty string not working
+- default ->group() class is now set to `c-form`
 
 ## [1.2.1] - 2022-09-06
 ### Fixed